backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
xgboost==1.7.6
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
python-dotenv==1.0.0
requests==2.31.0
